Importación de datos a un nuevo archivo para mejorar la organización del código
