+++ src/importarDatos.js
@@ -1,0 +1,13 @@
+const importarTodosLosDatos = () => {
+    const context = require.context('./datos/', false, /\.json$/);
+    const datos = {};
+  
+    context.keys().forEach(key => {
+      const nombreArchivo = key.replace('./', '').replace('.json', '');
+      datos[nombreArchivo] = context(key);
+    });
+    
+    return datos;
+  };
+
+export default importarTodosLosDatos;
